requirements.txt
Django==4.2.4
djangorestframework==3.14.0
django-filter==22.1